cast to avoid warning

+++ lib/init.c
@@ -139,7 +139,7 @@ usage( jname, justversion )
 
     fprintf(stderr, "\n");
     fprintf(stderr, "For memory region sizes a suffix of 'k' (1024) and 'm' (1024*1024)\n");
-    fprintf(stderr, "can be used. A descriptor takes up %d bytes.\n", DESCR );
+    fprintf(stderr, "can be used. A descriptor takes up %d bytes.\n", (int)DESCR );
     exit(1);
 }
